% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cellspbps}
\alias{cellspbps}
\title{select the cells to create pseudobulk pseudosamples.}
\usage{
cellspbps(subsample, pb_id, cell_id, seed = 1, n = 1)
}
\arguments{
\item{subsample}{A \code{data.frame} or \code{tibble} containing single-cell data with columns for group, bsg (biological subgroup), sample identifiers and cell identifiers.}

\item{pb_id}{A character string specifying the column name for biological replicate/sample identifier. Default is \code{"sample_cell"}.}

\item{cell_id}{A character string specifying the column name for cell identifiers. Default is \code{"cell_id"}.}

\item{seed}{A numeric value to set the random seed for reproducibility. Default is 1.}

\item{n}{An integer specifying how many pseudobulk pseudosamples to generate from each group. Default is 1.}
}
\value{
A list of character vectors containing sampled cell IDs for each pseudobulk pseudosample.
}
\description{
\code{cellspbps()} samples with replacement sets of cells from groups that share the same biological and technical conditions.
Preserving the average number of cells per biological subgroup.
}
\keyword{internal}
