% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_pbps.R
\name{gen_PBPS}
\alias{gen_PBPS}
\title{Generate a pseudobulk dataset with pseudobulk pseudosamples to be used as negative control samples in the removal of unwanted variation}
\usage{
gen_PBPS(
  ds,
  ctype = "cg_cov",
  BioVar,
  NVar,
  id_pb = "sample_cell",
  id_sub = "ind_cov",
  cell_id = "cell_id",
  n = 2,
  Seed = 2
)
}
\arguments{
\item{ds}{A Seurat object containing the single-cell data.}

\item{ctype}{A character string specifying the column in \code{ds@meta.data} used to define the cell type.}

\item{BioVar}{A character vector of column names in \code{ds@meta.data} representing biological variables (e.g., treatment).}

\item{NVar}{A character vector of column names in \code{ds@meta.data} representing technical/nuisance variables (e.g., batch ID).}

\item{id_pb}{Name of the column to use as sample ID (e.g., "sample_id").}

\item{id_sub}{Name of the column to store the "subject-level" ID (e.g., patient).}

\item{cell_id}{A character string indicating the column in \code{ds@meta.data} with the cell ID.}

\item{n}{Number of pseudo-replicates to generate per group. Default is 2.}

\item{Seed}{Integer used to set the random seed for reproducibility.}
}
\value{
A \code{DGEList} object (from edgeR) containing counts and metadata.
}
\description{
This function orchestrates the process of creating pseudobulk pseudosamples (PBPS) from a Seurat object and associated metadata.
It combines sampling, aggregation, and metadata reconstruction, and returns an edgeR-compatible \code{DGEList} object.
}
